requirements.txt
livekit-agents[deepgram,openai,elevenlabs,silero,turn-detector]~=1.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
elevenlabs>=0.2.0
deepgram-sdk>=2.12.0
silero-vad>=0.1.0 